Sectionify ES6ify navigation.
