+++ client/my-sites/theme/theme-sheet-content/index.jsx
@@ -10,9 +10,7 @@ import React from 'react';
  */
 import Gridicon from 'components/gridicon';
 import { localize } from 'i18n-calypso';
-import SectionNav from 'components/section-nav';
-import NavTabs from 'components/section-nav/tabs';
-import NavItem from 'components/section-nav/item';
+import Navigation from './theme-content-sections/navigation';
 import Overview from './theme-content-sections/overview';
 import Setup from './theme-content-sections/setup';
 import Support from './theme-content-sections/support';
@@ -63,52 +61,6 @@ class ThemeSheetContent extends React.Component {
 		theme: React.PropTypes.object,
 	};
 
-	getValidSections() {
-		const { theme } = this.props;
-		const validSections = [];
-		validSections.push( '' ); // Default section
-		theme && theme.supportDocumentation && validSections.push( 'setup' );
-		validSections.push( 'support' );
-		return validSections;
-	}
-
-	validateSection( section ) {
-		if ( this.getValidSections().indexOf( section ) === -1 ) {
-			return this.getValidSections()[ 0 ];
-		}
-		return section;
-	}
-
-	renderSectionNav( currentSection ) {
-		const { translate, siteSlug, id, isLoaded } = this.props;
-
-		const filterStrings = {
-			'': translate( 'Overview', { context: 'Filter label for theme content' } ),
-			setup: translate( 'Setup', { context: 'Filter label for theme content' } ),
-			support: translate( 'Support', { context: 'Filter label for theme content' } ),
-		};
-
-		const sitePart = siteSlug ? `/${ siteSlug }` : '';
-
-		const nav = (
-			<NavTabs label="Details" >
-				{ this.getValidSections().map( ( section ) => (
-					<NavItem key={ section }
-						path={ `/theme/${ id }${ section ? '/' + section : '' }${ sitePart }` }
-						selected={ section === currentSection }>
-						{ filterStrings[ section ] }
-					</NavItem>
-				) ) }
-			</NavTabs>
-		);
-
-		return (
-			<SectionNav className="theme__sheet-section-nav" selectedText={ filterStrings[ currentSection ] }>
-				{ isLoaded && nav }
-			</SectionNav>
-		);
-	}
-
 	renderSectionContent( section ) {
 		return {
 			'': <Overview { ...this.props } />,
@@ -118,14 +70,19 @@ class ThemeSheetContent extends React.Component {
 	}
 
 	render() {
-		const section = this.validateSection( this.props.section );
-		const { isJetpack, isLoaded, theme, togglePreview } = this.props;
+		const { id, isJetpack, isLoaded, theme, section, siteSlug } = this.props;
 
 		return (
 			<div className="theme__sheet-columns">
 				<div className="theme__sheet-column-left">
 					<div className="theme__sheet-content">
-						{ this.renderSectionNav( section ) }
+						<Navigation
+							siteSlug={ siteSlug }
+							id={ id }
+							isLoaded={ isLoaded }
+							currentSection={ section }
+							theme={ theme }
+						/>
 						{ this.renderSectionContent( section ) }
 						<div className="theme__sheet-footer-line"><Gridicon icon="my-sites" /></div>
 					</div>
@@ -135,7 +92,6 @@ class ThemeSheetContent extends React.Component {
 						isJetpack={ isJetpack }
 						isLoaded={ isLoaded }
 						theme={ theme }
-						togglePreview={ togglePreview }
 					/>
 				</div>
 			</div>
